add sync as sub-command
